split codegen.rs into multiple modules
