+++ src/ll/codegen.rs
@@ -1,213 +1,14 @@
 //! Bytecode generation.
 
-use std::{
-    collections::{HashMap, HashSet},
-    mem,
-    rc::Rc,
-};
+use std::{collections::HashSet, rc::Rc};
 
-use super::bytecode::{FunctionIndex, GlobalIndex, ImplementedTraitIndex, MethodIndex, TraitIndex};
+pub use self::traits::TraitBuilder;
+use self::{control_flow::BreakableBlock, structs::StructData, variables::Locals};
 use crate::ll::{
     ast::{Ast, NodeId, NodeKind},
-    bytecode::{
-        BuiltinTraits, CaptureKind, Chunk, Environment, Function, FunctionKind, MethodSignature,
-        Opcode, Opr24, Prototype,
-    },
-    common::{Error, ErrorKind, Location, RenderedSignature},
+    bytecode::{BuiltinTraits, Chunk, Environment, Opcode},
+    common::{Error, ErrorKind},
 };
-
-/// The index of a local on the stack.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[repr(transparent)]
-pub struct LocalIndex(Opr24);
-
-impl LocalIndex {
-    pub(crate) fn to_u32(self) -> u32 {
-        u32::from(self.0)
-    }
-}
-
-/// The index of an upvalue in a closure.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[repr(transparent)]
-pub struct UpvalueIndex(Opr24);
-
-impl UpvalueIndex {
-    pub(crate) fn to_u32(self) -> u32 {
-        u32::from(self.0)
-    }
-}
-
-/// Info about a local variable on the stack.
-#[derive(Debug)]
-struct Variable {
-    stack_slot: LocalIndex,
-    is_captured: bool,
-}
-
-#[derive(Debug, Default)]
-struct Scope {
-    /// Mapping from variable names to stack slots.
-    variables_by_name: HashMap<String, Variable>,
-    allocated_variable_count: u32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum VariablePlace {
-    Global(GlobalIndex),
-    Local(LocalIndex),
-    Upvalue(UpvalueIndex),
-}
-
-/// The kind of a variable allocation.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum VariableAllocation {
-    /// Inherit the allocation from the caller (parameter passing).
-    Inherit,
-    /// Allocate new storage for the variable.
-    Allocate,
-}
-
-/// Local variables, including upvalues.
-#[derive(Default)]
-struct Locals {
-    /// If there is a parent code generator with its own scopes (the current instance is in the
-    /// middle of compiling a closure), this is populated with its Locals.
-    parent: Option<Box<Self>>,
-
-    scopes: Vec<Scope>,
-    /// The next stack slot to be occupied by a variable.
-    local_count: u32,
-    /// The total amount of locals currently allocated. This is used to populate the
-    /// `preallocate_stack_slots` field in chunks, to provide more efficient allocations
-    allocated_local_count: u32,
-
-    /// Variables captured from parent scopes.
-    captures: Vec<CaptureKind>,
-}
-
-impl Locals {
-    /// Creates a new local.
-    fn create_local(
-        &mut self,
-        name: &str,
-        allocation: VariableAllocation,
-    ) -> Result<VariablePlace, ErrorKind> {
-        let slot = Opr24::new(self.local_count).map_err(|_| ErrorKind::TooManyLocals)?;
-        let slot = LocalIndex(slot);
-        let scope = self.scopes.last_mut().unwrap();
-        scope
-            .variables_by_name
-            .insert(name.to_owned(), Variable { stack_slot: slot, is_captured: false });
-        self.local_count += 1;
-        if allocation == VariableAllocation::Allocate {
-            self.allocated_local_count += 1;
-            scope.allocated_variable_count += 1;
-        }
-        Ok(VariablePlace::Local(slot))
-    }
-
-    fn variables_in_scope_mut(&mut self) -> impl Iterator<Item = (&str, &'_ mut Variable)> {
-        self.scopes
-            .iter_mut()
-            .rev()
-            .flat_map(|scope| scope.variables_by_name.iter_mut().map(|(s, v)| (s.as_str(), v)))
-    }
-
-    /// Returns the index of the given capture.
-    fn capture_index(&mut self, capture: CaptureKind) -> Result<UpvalueIndex, ErrorKind> {
-        // Iterating over captures maybe isn't most efficient here but it's not like we have
-        // thousands of them anyways. Unless somebody absolutely crazy starts writing Mica code.
-        // Then all I can say is: I hate you.
-        let index = self.captures.iter().rposition(|c| c.eq(&capture)).unwrap_or_else(|| {
-            let index = self.captures.len();
-            self.captures.push(capture);
-            index
-        });
-        Ok(UpvalueIndex(Opr24::try_from(index).map_err(|_| ErrorKind::TooManyCaptures)?))
-    }
-
-    /// Performs a local variable lookup. This may modify parent Locals and capture upvalues.
-    fn lookup(&mut self, name: &str) -> Result<Option<VariablePlace>, ErrorKind> {
-        // Work inside out: try innermost scopes (own locals) first.
-        for scope in self.scopes.iter().rev() {
-            if let Some(var) = scope.variables_by_name.get(name) {
-                return Ok(Some(VariablePlace::Local(var.stack_slot)));
-            }
-        }
-        // If there isn't a local with the given name, go up a level and look for locals to capture
-        // or existing upvalues.
-        if let Some(parent) = self.parent.as_mut() {
-            if let Some(place) = parent.lookup(name)? {
-                match place {
-                    VariablePlace::Local(local_slot) => {
-                        let (_, variable) = parent
-                            .variables_in_scope_mut()
-                            .find(|(_, var)| var.stack_slot == local_slot)
-                            .unwrap();
-                        variable.is_captured = true;
-                        let stack_slot = variable.stack_slot;
-                        let upvalue_index = self.capture_index(CaptureKind::Local(stack_slot))?;
-                        return Ok(Some(VariablePlace::Upvalue(upvalue_index)));
-                    }
-                    VariablePlace::Upvalue(upvalue_index) => {
-                        let own_index = self.capture_index(CaptureKind::Upvalue(upvalue_index))?;
-                        return Ok(Some(VariablePlace::Upvalue(own_index)));
-                    }
-                    VariablePlace::Global(_) => unreachable!(),
-                }
-            }
-        }
-        Ok(None)
-    }
-
-    /// Pushes a new scope onto the scope stack.
-    fn push_scope(&mut self) {
-        self.scopes.push(Default::default());
-    }
-
-    /// Pops the topmost scope off the scope stack and frees storage of any variables.
-    fn pop_scope(&mut self) -> Scope {
-        let scope = self.scopes.pop().expect("no scopes left on the stack");
-        self.local_count -= scope.variables_by_name.len() as u32;
-        self.allocated_local_count -= scope.allocated_variable_count;
-        scope
-    }
-}
-
-#[derive(Debug, Default)]
-struct BreakableBlock {
-    /// A list of offsets where `breaks` should be backpatched.
-    breaks: Vec<usize>,
-    start: usize,
-}
-
-#[derive(Debug, Default)]
-struct StructData {
-    /// Mapping from field names to indices.
-    fields: HashMap<Rc<str>, Opr24>,
-    /// The `self` variable. Used in `@field` lookups.
-    receiver: Option<VariablePlace>,
-}
-
-impl StructData {
-    /// Returns the index of the field with the given name, or creates that field if it doesn't
-    /// exist yet.
-    fn get_or_create_field(&mut self, name: &str) -> Result<Opr24, ErrorKind> {
-        if !self.fields.contains_key(name) {
-            let index = Opr24::try_from(self.fields.len()).map_err(|_| ErrorKind::TooManyFields)?;
-            self.fields.insert(Rc::from(name), index);
-            Ok(index)
-        } else {
-            Ok(*self.fields.get(name).unwrap())
-        }
-    }
-
-    /// Returns the index of the field with the given name, or `None` if there is no such field.
-    fn get_field(&self, name: &str) -> Option<Opr24> {
-        self.fields.get(name).copied()
-    }
-}
 
 pub struct CodeGenerator<'e> {
     env: &'e mut Environment,
@@ -246,97 +47,6 @@ impl<'e> CodeGenerator<'e> {
         }
     }
 
-    /// Creates a variable. If there is a scope on the stack, the variable is local; otherwise it
-    /// is global.
-    fn create_variable(
-        &mut self,
-        name: &str,
-        allocation: VariableAllocation,
-    ) -> Result<VariablePlace, ErrorKind> {
-        if !self.locals.scopes.is_empty() {
-            let place = self.locals.create_local(name, allocation)?;
-            self.chunk.preallocate_stack_slots =
-                self.chunk.preallocate_stack_slots.max(self.locals.allocated_local_count);
-            Ok(place)
-        } else {
-            let slot = self.env.create_global(name)?;
-            Ok(VariablePlace::Global(slot))
-        }
-    }
-
-    /// Performs a variable lookup. Returns the stack slot of the variable if it exists.
-    /// Otherwise returns `None`.
-    fn lookup_variable(&mut self, name: &str) -> Result<Option<VariablePlace>, ErrorKind> {
-        // Work from the inside out: check innermost local scopes first.
-        if let Some(place) = self.locals.lookup(name)? {
-            return Ok(Some(place));
-        }
-        // Lastly check globals.
-        Ok(self.env.get_global(name).map(VariablePlace::Global))
-    }
-
-    /// Pushes a new scope onto the scope stack.
-    fn push_scope(&mut self) {
-        self.locals.push_scope();
-    }
-
-    /// Pops the topmost scope off the scope stack and frees storage of any variables.
-    fn pop_scope(&mut self) {
-        let scope = self.locals.pop_scope();
-        for variable in scope.variables_by_name.into_values() {
-            if variable.is_captured {
-                self.chunk.emit((Opcode::CloseLocal, variable.stack_slot.0));
-            }
-        }
-    }
-
-    /// Generates a variable load instruction (GetLocal, GetGlobal, GetUpvalue).
-    fn generate_variable_load(&mut self, variable: VariablePlace) {
-        self.chunk.emit(match variable {
-            VariablePlace::Global(slot) => (Opcode::GetGlobal, slot.to_opr24()),
-            VariablePlace::Local(slot) => (Opcode::GetLocal, slot.0),
-            VariablePlace::Upvalue(slot) => (Opcode::GetUpvalue, slot.0),
-        });
-    }
-
-    /// Generates a variable assign instruction (AssignLocal, AssignGlobal, AssignUpvalue).
-    fn generate_variable_assign(&mut self, variable: VariablePlace) {
-        self.chunk.emit(match variable {
-            VariablePlace::Global(slot) => (Opcode::AssignGlobal, slot.to_opr24()),
-            VariablePlace::Local(slot) => (Opcode::AssignLocal, slot.0),
-            VariablePlace::Upvalue(slot) => (Opcode::AssignUpvalue, slot.0),
-        });
-    }
-
-    /// Generates a variable sink instruction (SinkLocal, SinkGlobal, SinkUpvalue).
-    fn generate_variable_sink(&mut self, variable: VariablePlace) {
-        self.chunk.emit(match variable {
-            VariablePlace::Global(slot) => (Opcode::SinkGlobal, slot.to_opr24()),
-            VariablePlace::Local(slot) => (Opcode::SinkLocal, slot.0),
-            VariablePlace::Upvalue(slot) => (Opcode::SinkUpvalue, slot.0),
-        });
-    }
-
-    /// Pushes a new breakable block.
-    fn push_breakable_block(&mut self) {
-        let start = self.chunk.emit(Opcode::Nop);
-        self.breakable_blocks.push(BreakableBlock { breaks: Vec::new(), start });
-    }
-
-    /// Pops the topmost breakable block.
-    fn pop_breakable_block(&mut self) {
-        let block = self.breakable_blocks.pop().unwrap();
-        if !block.breaks.is_empty() {
-            self.chunk.patch(block.start, Opcode::EnterBreakableBlock);
-            for jump in block.breaks {
-                // Unwrapping is safe here because if the loop is too large the error was caught
-                // already before `pop_breakable_block` was called.
-                self.chunk.patch(jump, Opcode::jump_forward(jump, self.chunk.len()).unwrap());
-            }
-            self.chunk.emit((Opcode::ExitBreakableBlock, 1));
-        }
-    }
-
     /// Generates code for a list of nodes. The last node's value is the one left on the stack.
     ///
     /// If there are no nodes in the list, this is equivalent to a `nil` literal.
@@ -353,898 +63,6 @@ impl<'e> CodeGenerator<'e> {
             }
         }
         Ok(())
-    }
-
-    /// Generates code for a nil literal.
-    fn generate_nil(&mut self) -> ExpressionResult {
-        self.chunk.emit(Opcode::PushNil);
-        ExpressionResult::Present
-    }
-
-    /// Generates code for a boolean literal.
-    fn generate_boolean(&mut self, ast: &Ast, node: NodeId) -> ExpressionResult {
-        self.chunk.emit(match ast.kind(node) {
-            NodeKind::True => Opcode::PushTrue,
-            NodeKind::False => Opcode::PushFalse,
-            _ => unreachable!(),
-        });
-        ExpressionResult::Present
-    }
-
-    /// Generates code for a number literal.
-    fn generate_number(&mut self, ast: &Ast, node: NodeId) -> ExpressionResult {
-        self.chunk.emit(Opcode::PushNumber);
-        let number = ast.number(node).unwrap();
-        self.chunk.emit_number(number);
-        ExpressionResult::Present
-    }
-
-    /// Generates code for a string literal.
-    fn generate_string(&mut self, ast: &Ast, node: NodeId) -> ExpressionResult {
-        self.chunk.emit(Opcode::PushString);
-        let string = ast.string(node).unwrap();
-        self.chunk.emit_string(string);
-        ExpressionResult::Present
-    }
-
-    /// Generates code for a unary operator.
-    fn generate_unary(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (left, _) = ast.node_pair(node);
-        self.generate_node(ast, left, Expression::Used)?;
-        match ast.kind(node) {
-            NodeKind::Negate => self.chunk.emit(Opcode::Negate),
-            NodeKind::Not => self.chunk.emit(Opcode::Not),
-            _ => unreachable!(),
-        };
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for a binary operator.
-    fn generate_binary(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (left, right) = ast.node_pair(node);
-        self.generate_node(ast, left, Expression::Used)?;
-        self.generate_node(ast, right, Expression::Used)?;
-        match ast.kind(node) {
-            NodeKind::Negate => self.chunk.emit(Opcode::Negate),
-
-            NodeKind::Add => self.chunk.emit(Opcode::Add),
-            NodeKind::Subtract => self.chunk.emit(Opcode::Subtract),
-            NodeKind::Multiply => self.chunk.emit(Opcode::Multiply),
-            NodeKind::Divide => self.chunk.emit(Opcode::Divide),
-
-            NodeKind::Equal => self.chunk.emit(Opcode::Equal),
-            NodeKind::NotEqual => {
-                self.chunk.emit(Opcode::Equal);
-                self.chunk.emit(Opcode::Not)
-            }
-            NodeKind::Less => self.chunk.emit(Opcode::Less),
-            NodeKind::LessEqual => self.chunk.emit(Opcode::LessEqual),
-            NodeKind::Greater => {
-                self.chunk.emit(Opcode::Swap);
-                self.chunk.emit(Opcode::Less)
-            }
-            NodeKind::GreaterEqual => {
-                self.chunk.emit(Opcode::Swap);
-                self.chunk.emit(Opcode::LessEqual)
-            }
-            _ => unreachable!(),
-        };
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for a variable lookup.
-    fn generate_variable(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let name = ast.string(node).unwrap();
-        if let Some(variable) = self.lookup_variable(name).map_err(|kind| ast.error(node, kind))? {
-            self.generate_variable_load(variable);
-            Ok(ExpressionResult::Present)
-        } else {
-            Err(ast.error(node, ErrorKind::VariableDoesNotExist(name.to_owned())))
-        }
-    }
-
-    /// Generates code for a field lookup.
-    fn generate_field(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (name, _) = ast.node_pair(node);
-        let name = ast.string(name).unwrap();
-        let struct_data = self
-            .struct_data
-            .as_deref()
-            .ok_or_else(|| ast.error(node, ErrorKind::FieldOutsideOfImpl))?;
-        let field_id = struct_data
-            .get_field(name)
-            .ok_or_else(|| ast.error(node, ErrorKind::FieldDoesNotExist(Rc::clone(name))))?;
-        // Unwrapping is OK here because fields are not allowed outside of functions, and each
-        // function with `StructData` passed in assigns to `receiver` at the start of its
-        // code generation.
-        let receiver = struct_data.receiver.unwrap();
-        self.generate_variable_load(receiver);
-        self.chunk.emit((Opcode::GetField, field_id));
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for destructuring a pattern.
-    ///
-    /// The generated code assumes there's a value to destructure at the top of the stack (called
-    /// the *scrutinee*.)
-    fn generate_pattern_destructuring(&mut self, ast: &Ast, node: NodeId) -> Result<(), Error> {
-        match ast.kind(node) {
-            NodeKind::Identifier => {
-                let name = ast.string(node).unwrap();
-                let variable = self
-                    .create_variable(name, VariableAllocation::Allocate)
-                    .map_err(|kind| ast.error(node, kind))?;
-                self.generate_variable_sink(variable);
-            }
-            _ => return Err(ast.error(node, ErrorKind::InvalidPattern)),
-        }
-        Ok(())
-    }
-
-    /// Generates code for an assignment.
-    fn generate_assignment(
-        &mut self,
-        ast: &Ast,
-        node: NodeId,
-        result: Expression,
-    ) -> Result<ExpressionResult, Error> {
-        let (target, value) = ast.node_pair(node);
-        self.generate_node(ast, value, Expression::Used)?;
-
-        match ast.kind(target) {
-            NodeKind::Identifier => {
-                let name = ast.string(target).unwrap();
-                let variable = if let Some(slot) =
-                    self.lookup_variable(name).map_err(|kind| ast.error(target, kind))?
-                {
-                    slot
-                } else {
-                    self.create_variable(name, VariableAllocation::Allocate)
-                        .map_err(|kind| ast.error(node, kind))?
-                };
-                match result {
-                    Expression::Used => self.generate_variable_assign(variable),
-                    Expression::Discarded => self.generate_variable_sink(variable),
-                }
-            }
-            NodeKind::Field => {
-                let (name, _) = ast.node_pair(target);
-                let name = ast.string(name).unwrap();
-                if let Some(struct_data) = self.struct_data.as_mut() {
-                    let field = if self.allow_new_fields {
-                        struct_data
-                            .get_or_create_field(name)
-                            .map_err(|kind| ast.error(node, kind))?
-                    } else {
-                        struct_data.get_field(name).ok_or_else(|| {
-                            ast.error(target, ErrorKind::FieldDoesNotExist(Rc::clone(name)))
-                        })?
-                    };
-                    // Unwrapping is OK here because `receiver` is assigned at the start of each
-                    // function in an `impl` block.
-                    let receiver = struct_data.receiver.unwrap();
-                    self.generate_variable_load(receiver);
-                    self.chunk.emit(match result {
-                        Expression::Used => (Opcode::AssignField, field),
-                        Expression::Discarded => (Opcode::SinkField, field),
-                    });
-                } else {
-                    return Err(ast.error(target, ErrorKind::FieldOutsideOfImpl));
-                }
-                // In constructors, we need to keep track of which fields were assigned to report
-                // errors about missing field values.
-                if self.is_constructor && !self.allow_new_fields {
-                    self.assigned_fields.insert(Rc::clone(name));
-                }
-            }
-            _ => return Err(ast.error(target, ErrorKind::InvalidAssignment)),
-        }
-
-        Ok(match result {
-            Expression::Discarded => ExpressionResult::Absent,
-            Expression::Used => ExpressionResult::Present,
-        })
-    }
-
-    /// Generates code for a list literal.
-    fn generate_list(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let children = ast.children(node).unwrap();
-        let len = Opr24::try_from(children.len())
-            .map_err(|_| ast.error(node, ErrorKind::ListIsTooLong))?;
-        for &child in children {
-            self.generate_node(ast, child, Expression::Used)?;
-        }
-        self.chunk.emit((Opcode::CreateList, len));
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for a dict literal.
-    fn generate_dict(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let children = ast.children(node).unwrap();
-        let len = Opr24::try_from(children.len())
-            .map_err(|_| ast.error(node, ErrorKind::DictIsTooLarge))?;
-        for &child in children {
-            let (key, value) = ast.node_pair(child);
-            self.generate_node(ast, key, Expression::Used)?;
-            self.generate_node(ast, value, Expression::Used)?;
-        }
-        self.chunk.emit((Opcode::CreateDict, len));
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for a `do..end` expression.
-    fn generate_do(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let children = ast.children(node).unwrap();
-        self.push_scope();
-        self.generate_node_list(ast, children)?;
-        self.pop_scope();
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for an `if..elif..else..end` expression.
-    fn generate_if(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let branches = ast.children(node).unwrap();
-        let mut jumps_to_end = Vec::new();
-
-        for (i, &branch) in branches.iter().enumerate() {
-            // We need to discard the previous branch's condition (if there was a previous branch).
-            if i > 0 {
-                self.chunk.emit(Opcode::Discard);
-            }
-
-            let then = ast.children(branch).unwrap();
-            match ast.kind(branch) {
-                NodeKind::IfBranch => {
-                    // Generate the condition.
-                    let (condition, _) = ast.node_pair(branch);
-                    self.push_scope();
-                    self.generate_node(ast, condition, Expression::Used)?;
-                    // Generate a Nop that is later backpatched with a ConditionalJumpForward.
-                    let jump = self.chunk.emit(Opcode::Nop);
-                    self.chunk.emit(Opcode::Discard); // The condition has to be discarded.
-                    self.generate_node_list(ast, then)?;
-                    self.pop_scope();
-                    let jump_to_end = self.chunk.emit(Opcode::Nop);
-                    jumps_to_end.push(jump_to_end);
-                    self.chunk.patch(
-                        jump,
-                        Opcode::jump_forward_if_falsy(jump, self.chunk.len())
-                            .map_err(|_| ast.error(branch, ErrorKind::IfBranchTooLarge))?,
-                    );
-                }
-
-                NodeKind::ElseBranch => {
-                    self.push_scope();
-                    self.generate_node_list(ast, then)?;
-                    self.pop_scope();
-                }
-
-                _ => unreachable!(),
-            }
-        }
-
-        // If there was no `else` branch, we need to patch in an implicit one that returns `nil`.
-        if ast.kind(*branches.last().unwrap()) != NodeKind::ElseBranch {
-            self.chunk.emit(Opcode::Discard);
-            self.chunk.emit(Opcode::PushNil);
-        }
-
-        // Backpatch all jumps to end with an unconditional jump forward.
-        for jump in jumps_to_end {
-            self.chunk.patch(
-                jump,
-                Opcode::jump_forward(jump, self.chunk.len())
-                    .map_err(|_| ast.error(node, ErrorKind::IfExpressionTooLarge))?,
-            );
-        }
-
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for an `and` infix operator.
-    fn generate_and(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (left, right) = ast.node_pair(node);
-        self.push_scope();
-        self.generate_node(ast, left, Expression::Used)?;
-        let jump_past_right = self.chunk.emit(Opcode::Nop);
-        self.chunk.emit(Opcode::Discard);
-        self.generate_node(ast, right, Expression::Used)?;
-        self.chunk.patch(
-            jump_past_right,
-            Opcode::jump_forward_if_falsy(jump_past_right, self.chunk.len())
-                .map_err(|_| ast.error(node, ErrorKind::OperatorRhsTooLarge))?,
-        );
-        self.pop_scope();
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for an `or` infix operator.
-    fn generate_or(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (left, right) = ast.node_pair(node);
-        self.push_scope();
-        self.generate_node(ast, left, Expression::Used)?;
-        let jump_past_right = self.chunk.emit(Opcode::Nop);
-        self.chunk.emit(Opcode::Discard);
-        self.generate_node(ast, right, Expression::Used)?;
-        self.chunk.patch(
-            jump_past_right,
-            Opcode::jump_forward_if_truthy(jump_past_right, self.chunk.len())
-                .map_err(|_| ast.error(node, ErrorKind::OperatorRhsTooLarge))?,
-        );
-        self.pop_scope();
-        Ok(ExpressionResult::Present)
-    }
-
-    fn generate_conditional_loop(
-        &mut self,
-        ast: &Ast,
-        node: NodeId,
-        generate_condition: &dyn Fn(&mut CodeGenerator<'_>) -> Result<(), Error>,
-        generate_body: &dyn Fn(&mut CodeGenerator<'_>) -> Result<(), Error>,
-    ) -> Result<(), Error> {
-        // The outer scope, so that variables can be declared in the condition.
-        self.push_scope();
-        // The breakable block.
-        self.push_breakable_block();
-
-        let start = self.chunk.len();
-        generate_condition(self)?;
-        let jump_to_end = self.chunk.emit(Opcode::Nop);
-        // Discard the condition if it's true.
-        self.chunk.emit(Opcode::Discard);
-
-        generate_body(self)?;
-        // While loops don't yield a value.
-        self.chunk.emit(Opcode::Discard);
-
-        self.chunk.emit(
-            Opcode::jump_backward(self.chunk.len(), start)
-                .map_err(|_| ast.error(node, ErrorKind::LoopTooLarge))?,
-        );
-        self.chunk.patch(
-            jump_to_end,
-            Opcode::jump_forward_if_falsy(jump_to_end, self.chunk.len())
-                .map_err(|_| ast.error(node, ErrorKind::LoopTooLarge))?,
-        );
-        // Discard the condition if it's false.
-        self.chunk.emit(Opcode::Discard);
-
-        // Because loops are expressions, they must produce a value. That value is `nil`.
-        self.chunk.emit(Opcode::PushNil);
-
-        // `break`s produce a value (or `nil` by default), so we need to jump over the
-        // fallback `PushNil`.
-        self.pop_breakable_block();
-        self.pop_scope();
-
-        Ok(())
-    }
-
-    /// Generates code for a `while..do..end` loop.
-    fn generate_while(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (condition, _) = ast.node_pair(node);
-        let body = ast.children(node).unwrap();
-
-        self.generate_conditional_loop(
-            ast,
-            node,
-            &|generator| generator.generate_node(ast, condition, Expression::Used),
-            &|generator| generator.generate_node_list(ast, body),
-        )?;
-
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for a `for` loop.
-    fn generate_for(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (binding, iterator) = ast.node_pair(node);
-        let body = ast.children(node).unwrap();
-
-        self.push_scope();
-
-        let iterator_var = self
-            .create_variable("<iterator>", VariableAllocation::Allocate)
-            .map_err(|kind| ast.error(iterator, kind))?;
-        self.generate_node(ast, iterator, Expression::Used)?;
-        self.generate_variable_sink(iterator_var);
-
-        self.generate_conditional_loop(
-            ast,
-            node,
-            &|generator| {
-                generator.generate_variable_load(iterator_var);
-                generator.chunk.emit((
-                    Opcode::CallMethod,
-                    Opr24::pack((generator.builtin_traits.iterator_has_next.to_u16(), 1)),
-                ));
-                Ok(())
-            },
-            &|generator| {
-                generator.generate_variable_load(iterator_var);
-                generator.chunk.emit((
-                    Opcode::CallMethod,
-                    Opr24::pack((generator.builtin_traits.iterator_next.to_u16(), 1)),
-                ));
-                generator.generate_pattern_destructuring(ast, binding)?;
-                generator.generate_node_list(ast, body)?;
-                Ok(())
-            },
-        )?;
-
-        self.pop_scope();
-
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates a `break` expression.
-    fn generate_break(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (right, _) = ast.node_pair(node);
-        if right != NodeId::EMPTY {
-            self.generate_node(ast, right, Expression::Used)?;
-        } else {
-            let _ = self.generate_nil();
-        }
-        let jump = self.chunk.emit(Opcode::Nop);
-        if let Some(block) = self.breakable_blocks.last_mut() {
-            block.breaks.push(jump);
-        } else {
-            return Err(ast.error(node, ErrorKind::BreakOutsideOfLoop));
-        }
-        Ok(ExpressionResult::NoReturn)
-    }
-
-    /// Generates code for a function call.
-    fn generate_call(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (function, _) = ast.node_pair(node);
-        match ast.kind(function) {
-            // Method calls need special treatment.
-            NodeKind::Dot => {
-                let (receiver, name) = ast.node_pair(function);
-                if ast.kind(name) != NodeKind::Identifier {
-                    return Err(ast.error(name, ErrorKind::InvalidMethodName));
-                }
-                let name = ast.string(name).unwrap();
-
-                // Generate the receiver and arguments.
-                self.generate_node(ast, receiver, Expression::Used)?;
-                let arguments = ast.children(node).unwrap();
-                for &argument in arguments {
-                    self.generate_node(ast, argument, Expression::Used)?;
-                }
-
-                // Construct the call.
-                let arity: u8 = (1 + arguments.len())
-                    // ↑ Add 1 for the receiver, which is also an argument.
-                    .try_into()
-                    .map_err(|_| ast.error(node, ErrorKind::TooManyArguments))?;
-                let signature = MethodSignature::new(Rc::clone(name), arity as u16);
-                let method_index = self
-                    .env
-                    .get_or_create_method_index(&signature)
-                    .map_err(|kind| ast.error(node, kind))?;
-                self.chunk.emit((Opcode::CallMethod, Opr24::pack((method_index.to_u16(), arity))));
-            }
-            _ => {
-                self.generate_node(ast, function, Expression::Used)?;
-                let arguments = ast.children(node).unwrap();
-                for &argument in arguments {
-                    self.generate_node(ast, argument, Expression::Used)?;
-                }
-                self.chunk.emit((
-                    Opcode::Call,
-                    Opr24::try_from(arguments.len())
-                        .map_err(|_| ast.error(node, ErrorKind::TooManyArguments))?,
-                ));
-            }
-        }
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for a bare dot call (without parentheses).
-    fn generate_dot(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (left, method) = ast.node_pair(node);
-        self.generate_node(ast, left, Expression::Used)?;
-
-        if ast.kind(method) != NodeKind::Identifier {
-            return Err(ast.error(method, ErrorKind::InvalidMethodName));
-        }
-        let signature = MethodSignature::new(Rc::clone(ast.string(method).unwrap()), 1);
-        let method_index = self
-            .env
-            .get_or_create_method_index(&signature)
-            .map_err(|kind| ast.error(node, kind))?;
-        self.chunk.emit((Opcode::CallMethod, Opr24::pack((method_index.to_u16(), 1))));
-
-        Ok(ExpressionResult::Present)
-    }
-
-    fn generate_function(
-        &mut self,
-        ast: &Ast,
-        node: NodeId,
-        GenerateFunctionOptions { name, call_conv }: GenerateFunctionOptions,
-    ) -> Result<GeneratedFunction, Error> {
-        let (head, body) = ast.node_pair(node);
-        let (_, parameters) = ast.node_pair(head);
-        let parameter_list = ast.children(parameters).unwrap();
-
-        let mut generator =
-            CodeGenerator::new(Rc::clone(&self.chunk.module_name), self.env, self.builtin_traits);
-        // NOTE: Hopefully the allocation from this mem::take gets optimized out.
-        generator.locals.parent = Some(mem::take(&mut self.locals));
-        if call_conv.has_field_access() {
-            generator.struct_data = self.struct_data.take();
-            generator.allow_new_fields = call_conv.allow_new_fields();
-        }
-        generator.is_constructor = call_conv.is_constructor();
-
-        // Push a scope to start creating local variables.
-        generator.push_scope();
-        // Create local variables for all the parameters.
-        // Note that in bare functions, the function itself acts as the `self` parameter.
-        let receiver = if call_conv.has_visible_self() {
-            let receiver = generator.create_variable("self", VariableAllocation::Inherit).unwrap();
-            if let Some(struct_data) = generator.struct_data.as_deref_mut() {
-                struct_data.receiver = Some(receiver);
-            }
-            receiver
-        } else {
-            generator.create_variable("<receiver>", VariableAllocation::Inherit).unwrap()
-        };
-        for &parameter in parameter_list {
-            let parameter_name = ast.string(parameter).unwrap();
-            generator
-                .create_variable(parameter_name, VariableAllocation::Inherit)
-                .map_err(|kind| ast.error(parameter, kind))?;
-        }
-
-        // In constructors, we have to create `self` explicitly.
-        let create_struct = if call_conv.is_constructor() {
-            generator.generate_variable_load(receiver);
-            let instruction = generator.chunk.emit(Opcode::Nop);
-            let receiver = generator
-                .create_variable("self", VariableAllocation::Allocate)
-                .map_err(|kind| ast.error(node, kind))?;
-            generator.generate_variable_assign(receiver);
-            generator.chunk.emit(Opcode::Discard);
-            generator.struct_data.as_deref_mut().unwrap().receiver = Some(receiver);
-            Some(instruction)
-        } else {
-            None
-        };
-
-        // Generate the body.
-        generator.generate_node(ast, body, Expression::Used)?;
-
-        // If we're in a constructor we have to backpatch the `CreateStruct` instruction with the
-        // number of fields.
-        if let Some(create_struct) = create_struct {
-            let field_count = generator.struct_data.as_ref().unwrap().fields.len();
-            let field_count = Opr24::try_from(field_count)
-                .map_err(|_| ast.error(ast.node_pair(parameters).0, ErrorKind::TooManyFields))?;
-            generator.chunk.patch(create_struct, (Opcode::CreateStruct, field_count));
-            // We also have to discard whatever was at the top of the stack at the moment and
-            // return the struct we constructed.
-            generator.chunk.emit(Opcode::Discard);
-            let receiver = generator.struct_data.as_ref().unwrap().receiver.unwrap();
-            generator.generate_variable_load(receiver);
-        }
-        // If we're in a constructor, we also have to check if all fields were assigned.
-        if call_conv.is_constructor() && !call_conv.allow_new_fields() {
-            let struct_data = generator.struct_data.as_ref().unwrap();
-            let missing: Vec<_> = struct_data
-                .fields
-                .keys()
-                .filter(|field| !generator.assigned_fields.contains(*field))
-                .cloned()
-                .collect();
-            if !missing.is_empty() {
-                return Err(ast.error(node, ErrorKind::MissingFields(missing)));
-            }
-        }
-
-        // Finish generating the chunk by inserting a `Return` opcode.
-        generator.pop_scope();
-        generator.chunk.emit(Opcode::Return);
-
-        // Take back what was taken from the parent generator.
-        self.locals = generator.locals.parent.take().unwrap();
-        if call_conv.has_field_access() {
-            self.struct_data = generator.struct_data;
-        }
-
-        // Construct the function.
-        let parameter_count = u16::try_from(parameter_list.len())
-            .map_err(|_| ast.error(parameters, ErrorKind::TooManyParameters))?;
-        let function = Function {
-            name,
-            parameter_count: Some(parameter_count),
-            kind: FunctionKind::Bytecode {
-                chunk: Rc::new(generator.chunk),
-                captured_locals: generator.locals.captures,
-            },
-            hidden_in_stack_traces: false,
-        };
-        let function_id =
-            self.env.create_function(function).map_err(|kind| ast.error(node, kind))?;
-
-        Ok(GeneratedFunction { id: function_id, parameter_count })
-    }
-
-    /// Ensures that a `Func` node is a valid bare function - without a kind, and with a body.
-    fn ensure_valid_bare_function(ast: &Ast, node: NodeId) -> Result<(), Error> {
-        let (head, body) = ast.node_pair(node);
-        let (_name, parameters) = ast.node_pair(head);
-
-        let function_kind = ast.node_pair(parameters).0;
-        if function_kind != NodeId::EMPTY {
-            return Err(ast.error(function_kind, ErrorKind::FunctionKindOutsideImpl));
-        }
-        if body == NodeId::EMPTY {
-            return Err(ast.error(node, ErrorKind::MissingFunctionBody));
-        }
-        Ok(())
-    }
-
-    /// Generates code for a bare function declaration.
-    fn generate_function_declaration(
-        &mut self,
-        ast: &Ast,
-        node: NodeId,
-    ) -> Result<ExpressionResult, Error> {
-        Self::ensure_valid_bare_function(ast, node)?;
-
-        let (head, _) = ast.node_pair(node);
-        let (name_node, _) = ast.node_pair(head);
-        let name = ast.string(name_node).unwrap();
-
-        // Create the variable before compiling the function, to allow for recursion.
-        let variable = self
-            .create_variable(name, VariableAllocation::Allocate)
-            .map_err(|kind| ast.error(name_node, kind))?;
-
-        let function = self.generate_function(
-            ast,
-            node,
-            GenerateFunctionOptions { name: Rc::clone(name), call_conv: FunctionCallConv::Bare },
-        )?;
-
-        self.chunk.emit((Opcode::CreateClosure, function.id.to_opr24()));
-        self.generate_variable_sink(variable);
-
-        Ok(ExpressionResult::Absent)
-    }
-
-    /// Generates code for a function expression.
-    fn generate_function_expression(
-        &mut self,
-        ast: &Ast,
-        node: NodeId,
-    ) -> Result<ExpressionResult, Error> {
-        Self::ensure_valid_bare_function(ast, node)?;
-
-        let function = self.generate_function(
-            ast,
-            node,
-            GenerateFunctionOptions {
-                name: Rc::from("<anonymous>"),
-                call_conv: FunctionCallConv::Bare,
-            },
-        )?;
-
-        self.chunk.emit((Opcode::CreateClosure, function.id.to_opr24()));
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for a `return` expression.
-    fn generate_return(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (value, _) = ast.node_pair(node);
-        // Unlike breaking out of a loop, returning is super simple, as it implies that all locals
-        // are taken off the stack.
-        if value == NodeId::EMPTY {
-            let _ = self.generate_nil();
-        } else {
-            self.generate_node(ast, value, Expression::Used)?;
-        }
-        self.chunk.emit(Opcode::Return);
-        Ok(ExpressionResult::NoReturn)
-    }
-
-    /// Generates code for a struct declaration.
-    fn generate_struct(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (name, _) = ast.node_pair(node);
-        let name = ast.string(name).unwrap();
-
-        self.chunk.emit(Opcode::CreateType);
-        self.chunk.emit_string(name);
-        let variable = self
-            .create_variable(name, VariableAllocation::Allocate)
-            .map_err(|kind| ast.error(node, kind))?;
-        self.generate_variable_assign(variable);
-
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for an `impl` block.
-    fn generate_impl(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (implementee, _) = ast.node_pair(node);
-        let items = ast.children(node).unwrap();
-
-        // Push the implementee onto the stack first. Implemented traits follow.
-        self.generate_node(ast, implementee, Expression::Used)?;
-
-        // There's no need to save any old struct data because `impl` blocks don't nest freely.
-        // Yes, you can create an `impl` block in a function inside this `impl` block, but that
-        // function is handled by a different generator.
-        self.struct_data = Some(Box::new(StructData::default()));
-
-        let mut proto = Prototype::default();
-        let mut state = ImplGenerationState {
-            proto: &mut proto,
-            has_constructor: false,
-            implemented_trait_index: None,
-        };
-        for &node in items {
-            self.generate_impl_item(ast, node, &mut state, true)?;
-        }
-
-        let proto_id = self.env.create_prototype(proto).map_err(|kind| ast.error(node, kind))?;
-        self.chunk.emit((Opcode::Implement, proto_id.to_opr24()));
-
-        self.struct_data = None;
-
-        Ok(ExpressionResult::Present)
-    }
-
-    /// Generates code for a single item in an `impl` block.
-    fn generate_impl_item(
-        &mut self,
-        ast: &Ast,
-        node: NodeId,
-        state: &mut ImplGenerationState,
-        allow_as: bool,
-    ) -> Result<(), Error> {
-        match ast.kind(node) {
-            NodeKind::Func => {
-                let (head, _) = ast.node_pair(node);
-                let (name_node, parameters) = ast.node_pair(head);
-                if ast.kind(name_node) != NodeKind::Identifier {
-                    return Err(ast.error(node, ErrorKind::MissingMethodName));
-                }
-                let name = Rc::clone(ast.string(name_node).unwrap());
-                let (kind, _) = ast.node_pair(parameters);
-
-                let call_conv = match ast.kind(kind) {
-                    NodeKind::Empty => FunctionCallConv::Instance,
-                    NodeKind::Static => FunctionCallConv::Static,
-                    NodeKind::Constructor => {
-                        let allow_new_fields = !state.has_constructor;
-                        state.has_constructor = true;
-                        FunctionCallConv::Constructor { allow_new_fields }
-                    }
-                    _ => unreachable!(),
-                };
-                let function = self.generate_function(
-                    ast,
-                    node,
-                    GenerateFunctionOptions { name: Rc::clone(&name), call_conv },
-                )?;
-
-                if let Some(trait_index) = state.implemented_trait_index {
-                    // For trait instance methods, method ID resolution is performed at runtime.
-                    let key = (Rc::clone(&name), 1 + function.parameter_count, trait_index);
-                    if state.proto.trait_instance.insert(key, function.id).is_some() {
-                        return Err(ast.error(
-                            name_node,
-                            ErrorKind::MethodAlreadyImplemented(RenderedSignature {
-                                name,
-                                arity: Some(function.parameter_count),
-                                // Note that we do not have a way of knowing the actual name of the
-                                // trait, so we simply don't display
-                                // it.
-                                trait_name: None,
-                            }),
-                        ));
-                    }
-                } else {
-                    // For regular instance methods, method ID resolution is performed during
-                    // compilation.
-                    let signature = MethodSignature::new(name, 1 + function.parameter_count);
-                    let method_id = self
-                        .env
-                        .get_or_create_method_index(&signature)
-                        .map_err(|kind| ast.error(node, kind))?;
-
-                    let map = match ast.kind(kind) {
-                        NodeKind::Empty => &mut state.proto.instance,
-                        NodeKind::Static | NodeKind::Constructor => &mut state.proto.statics,
-                        _ => unreachable!(),
-                    };
-                    if map.insert(method_id, function.id).is_some() {
-                        return Err(ast.error(
-                            name_node,
-                            ErrorKind::MethodAlreadyImplemented(RenderedSignature {
-                                name: signature.name,
-                                arity: Some(function.parameter_count),
-                                trait_name: None,
-                            }),
-                        ));
-                    }
-                }
-            }
-
-            NodeKind::ImplAs => {
-                if !allow_as {
-                    return Err(ast.error(node, ErrorKind::AsCannotNest));
-                }
-                let (trait_expr, _) = ast.node_pair(node);
-                let as_items = ast.children(node).unwrap();
-                self.generate_node(ast, trait_expr, Expression::Used)?;
-                let trait_index =
-                    state.proto.implement_next_trait().map_err(|e| ast.error(node, e))?;
-                let mut state = ImplGenerationState {
-                    proto: state.proto,
-                    implemented_trait_index: Some(trait_index),
-                    ..*state
-                };
-                for &item in as_items {
-                    self.generate_impl_item(ast, item, &mut state, false)?;
-                }
-            }
-
-            // NB: If other item types are allowed, don't forget to change the error message for
-            // InvalidImplItem.
-            _ => return Err(ast.error(node, ErrorKind::InvalidImplItem)),
-        }
-
-        Ok(())
-    }
-
-    /// Generates code for a trait definition.
-    fn generate_trait(&mut self, ast: &Ast, node: NodeId) -> Result<ExpressionResult, Error> {
-        let (trait_name, _) = ast.node_pair(node);
-        let trait_name = ast.string(trait_name).unwrap();
-        let items = ast.children(node).unwrap();
-
-        let mut builder = TraitBuilder::new(self.env, Some(&self.chunk), Rc::clone(trait_name))
-            .map_err(|e| ast.error(node, e))?;
-
-        for &item in items {
-            match ast.kind(item) {
-                NodeKind::Func => {
-                    let (head, body) = ast.node_pair(item);
-                    let (name, params) = ast.node_pair(head);
-                    if body != NodeId::EMPTY {
-                        return Err(ast.error(body, ErrorKind::TraitMethodCannotHaveBody));
-                    }
-                    if name == NodeId::EMPTY {
-                        return Err(ast.error(head, ErrorKind::MissingMethodName));
-                    }
-                    if ast.node_pair(params).0 != NodeId::EMPTY {
-                        return Err(ast.error(head, ErrorKind::FunctionKindInTrait));
-                    }
-
-                    let _method_id = builder
-                        .add_method(
-                            Rc::clone(ast.string(name).unwrap()),
-                            u16::try_from(ast.len(params).unwrap())
-                                .map_err(|_| ast.error(item, ErrorKind::TooManyParameters))?,
-                        )
-                        .map_err(|e| ast.error(item, e))?;
-                }
-                _ => return Err(ast.error(item, ErrorKind::InvalidTraitItem)),
-            }
-        }
-
-        let (trait_id, _) = builder.build();
-
-        self.chunk.emit((Opcode::CreateTrait, trait_id.to_opr24()));
-        let variable = self
-            .create_variable(trait_name, VariableAllocation::Allocate)
-            .map_err(|k| ast.error(node, k))?;
-        self.generate_variable_assign(variable);
-
-        Ok(ExpressionResult::Present)
     }
 
     /// Generates code for a valid expression node.
@@ -1345,50 +163,6 @@ impl<'e> CodeGenerator<'e> {
     }
 }
 
-/// The calling convention of a function.
-#[derive(Debug, Clone, Copy)]
-enum FunctionCallConv {
-    /// A bare function: `self` is invisible, fields cannot be assigned.
-    Bare,
-    /// A static function: `self` is visible, fields cannot be assigned.
-    Static,
-    /// An instance function: `self` is visible, fields can be assigned but not created.
-    Instance,
-    /// A constructor: `self` is created from scratch, fields can created if `allow_new_fields` is
-    /// true.
-    Constructor { allow_new_fields: bool },
-}
-
-impl FunctionCallConv {
-    fn has_field_access(&self) -> bool {
-        matches!(self, Self::Instance | Self::Constructor { .. })
-    }
-
-    fn has_visible_self(&self) -> bool {
-        // Note that `Constructor` does not have a visible `self`, because it's an allocated local
-        // rather than inherited.
-        matches!(self, Self::Static | Self::Instance)
-    }
-
-    fn is_constructor(&self) -> bool {
-        matches!(self, Self::Constructor { .. })
-    }
-
-    fn allow_new_fields(&self) -> bool {
-        matches!(self, Self::Constructor { allow_new_fields: true })
-    }
-}
-
-struct GenerateFunctionOptions {
-    name: Rc<str>,
-    call_conv: FunctionCallConv,
-}
-
-struct GeneratedFunction {
-    id: FunctionIndex,
-    parameter_count: u16,
-}
-
 /// What kind of expression to generate in this place.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 enum Expression {
@@ -1409,104 +183,15 @@ enum ExpressionResult {
     NoReturn,
 }
 
-/// The state of generating an `impl` block.
-struct ImplGenerationState<'p> {
-    proto: &'p mut Prototype,
-    has_constructor: bool,
-    implemented_trait_index: Option<ImplementedTraitIndex>,
-}
-
-pub struct TraitBuilder<'b> {
-    pub env: &'b mut Environment,
-    parent_chunk: Option<&'b Chunk>,
-    trait_id: TraitIndex,
-    required_methods: HashSet<MethodIndex>,
-    shims: Vec<(MethodIndex, FunctionIndex)>,
-}
-
-impl<'b> TraitBuilder<'b> {
-    pub fn new(
-        env: &'b mut Environment,
-        parent_chunk: Option<&'b Chunk>,
-        name: Rc<str>,
-    ) -> Result<Self, ErrorKind> {
-        let trait_id = env.create_trait(name)?;
-        Ok(Self { env, parent_chunk, trait_id, required_methods: HashSet::new(), shims: vec![] })
-    }
-
-    /// Generates the *shim* for a trait method. The shim is responsible for calling the actual
-    /// method.
-    fn generate_method_shim(
-        &mut self,
-        trait_name: &str,
-        method_id: MethodIndex,
-        method_signature: &MethodSignature,
-    ) -> Result<FunctionIndex, ErrorKind> {
-        let arity = method_signature.arity.unwrap() as u8;
-
-        let (module_name, codegen_location) = if let Some(parent_chunk) = self.parent_chunk {
-            (Rc::clone(&parent_chunk.module_name), parent_chunk.codegen_location)
-        } else {
-            (Rc::from("FFI"), Location::UNINIT)
-        };
-
-        let mut chunk = Chunk::new(module_name);
-        chunk.codegen_location = codegen_location;
-        chunk.emit((Opcode::CallMethod, Opr24::pack((method_id.to_u16(), arity as u8))));
-        chunk.emit(Opcode::Return);
-
-        let shim_name = Rc::from(format!("trait {trait_name}.{} <shim>", method_signature.name));
-        let chunk = Rc::new(chunk);
-        let function_id = self.env.create_function(Function {
-            name: shim_name,
-            parameter_count: method_signature.arity,
-            kind: FunctionKind::Bytecode { chunk, captured_locals: vec![] },
-            hidden_in_stack_traces: true,
-        })?;
-
-        Ok(function_id)
-    }
-
-    /// Adds a function into the trait and returns its method ID and function ID.
-    ///
-    /// The arity does not include the `self` parameter.
-    pub fn add_method(&mut self, name: Rc<str>, arity: u16) -> Result<MethodIndex, ErrorKind> {
-        let trait_name = Rc::clone(&self.env.get_trait(self.trait_id).unwrap().name);
-        let signature = MethodSignature {
-            name,
-            arity: Some(arity.checked_add(1).ok_or(ErrorKind::TooManyParameters)?),
-            trait_id: Some(self.trait_id),
-        };
-        let method_id = self.env.get_or_create_method_index(&signature)?;
-
-        let shim_signature = MethodSignature {
-            // Add 2 for the receiving trait and self.
-            arity: Some(arity.checked_add(2).ok_or(ErrorKind::TooManyParameters)?),
-            trait_id: None,
-            ..signature.clone()
-        };
-        let shim_method_id = self.env.get_or_create_method_index(&shim_signature)?;
-        let shim_function_id = self.generate_method_shim(&trait_name, method_id, &signature)?;
-        self.shims.push((shim_method_id, shim_function_id));
-
-        if !self.required_methods.insert(method_id) {
-            return Err(ErrorKind::TraitAlreadyHasMethod(RenderedSignature {
-                name: signature.name,
-                arity: signature.arity,
-                // Do not duplicate the trait name in the error message.
-                trait_name: None,
-            }));
-        }
-
-        Ok(method_id)
-    }
-
-    /// Finishes building the trait, returns its trait ID, and gives back the mutable reference to
-    /// the environment.
-    pub fn build(self) -> (TraitIndex, &'b mut Environment) {
-        let prototype = self.env.get_trait_mut(self.trait_id).unwrap();
-        prototype.required = self.required_methods;
-        prototype.shims = self.shims;
-        (self.trait_id, self.env)
-    }
-}
+// The actual implementation is split into multiple files so as not to have to maintain a
+// multi-thousand line behemoth.
+mod assignment;
+mod calls;
+mod control_flow;
+mod functions;
+mod impls;
+mod literals;
+mod operators;
+mod structs;
+mod traits;
+pub mod variables;
